tests/requirements/base.txt
# requirements from setup.py
django-frontify
# other requirements
django-app-helper
tox
coverage